splits chart files and types data object
