+++ src/spotify/chart/config.ts
@@ -1,0 +1,1 @@
+export const config = {}
