+++ src/spotify/chart/index.ts
@@ -25,7 +25,9 @@ import {
   Tooltip
 } from 'chart.js'
 
-import { SongType } from './types'
+import { SongType } from '../types'
+import { config } from './config'
+import { data } from './setup'
 
 Chart.register(
   ArcElement,
@@ -53,39 +55,14 @@ Chart.register(
   Tooltip
 )
 
-export function chart(history: SongType[]) {
+export function chart() {
   const elem = document.getElementById('myChart') as HTMLCanvasElement
   const ctx = elem ? elem.getContext('2d') : null
 
   if (ctx)
     new Chart(ctx, {
       type: 'bar',
-      data: {
-        labels: ['Red', 'Blue', 'Yellow', 'Green', 'Purple', 'Orange'],
-        datasets: [
-          {
-            label: '# of Votes',
-            data: [12, 19, 3, 5, 2, 3],
-            backgroundColor: [
-              'rgba(255, 99, 132, 0.2)',
-              'rgba(54, 162, 235, 0.2)',
-              'rgba(255, 206, 86, 0.2)',
-              'rgba(75, 192, 192, 0.2)',
-              'rgba(153, 102, 255, 0.2)',
-              'rgba(255, 159, 64, 0.2)'
-            ],
-            borderColor: [
-              'rgba(255, 99, 132, 1)',
-              'rgba(54, 162, 235, 1)',
-              'rgba(255, 206, 86, 1)',
-              'rgba(75, 192, 192, 1)',
-              'rgba(153, 102, 255, 1)',
-              'rgba(255, 159, 64, 1)'
-            ],
-            borderWidth: 1
-          }
-        ]
-      },
+      data,
       options: {
         scales: {
           y: {
